3D Grid system overhaul + some small changes
Created a new 3d grid system using shaders instead of gDrawLine.

added a new class called gGrid to manage the grid and its states.
gGrid inherits gRenderObject and includes gVbo. one issue i faced was in the setvertexdata() method; there wasn't one that aligned with the shader requirements so I settled on the float array one because it only sets 1 attribute instead of 5 which was a problem because the grid holds the vertex data internally in the shader .vert file so to fix this i added the statement "in vec3 aPos;" inside the .vert file even though aPos is never used.

To implement gridshader variable/ use it, I followed the shaders management in gRenderer and  to implement class I took example of gCamera and gMesh design.

I migrated all customization options/functions from gRenderer and added new ones to the grid system like setting visibility distance using near and far clips of the camera.
Though currently they dont do anything different since the shader files doesnt support them fully.
Only near and far clips are available to customize and the shader is only for XZ grid for now.

made some minor changes to gCamera:
changed the near and far clip setters to clamp the values down so that its easier to control them in the canvas without the camera getting flipped and because the grid depends on them.
added near and far clip getters to gCamera.

other changes:
deleted all grid functions in the renderer and made a simple gDrawGrid() function that acts similar to the other gDraw functions.
deleted all grid functions from gBaseCanvas since they were dependent on the same functions in the renderer.
deleted the drawgrid call from gCamera .cpp
added shader_grid files to the engine directory for easier maintenance and development.
add gGrid.cpp to the engine's CMakeLists