folder display and icons added
A file management table has been created here and the necessary functions for its management have been added. The table is set to show the contents of the specified directory and icons are added to the files. The user interface component, such as returning to the previous directory, can also be managed.

The referencing method defined in gGUIButton -@bool& isPressedRef();- is used in the gGUITable class to keep track of the state of the previous button