compatibility with old engine version + fixes
put back the grid methods in base canvas and renderer then changed their implementation to work with gGrid class.

fixed semantic errors of Axis-related method signatures and variables and changed the draw methods to be more cohesive and draw the needed grid elements regardless of isenabled state, leaving that for the main draw method that gCamera::begin() indirectly calls.